docs/requirements.txt
pyyaml
pydata-sphinx-theme
